Integrated Value of Influence (IVI)

This function calculates the IVI of the desired nodes from previously calculated centrality measures. This function is not dependent to other packages and the required centrality measures, namely degree centrality, ClusterRank, betweenness centrality, Collective Influence, local H-index, and neighborhood connectivity could have been calculated by any means beforehand. A shiny app has also been developed for the calculation of IVI as well as IVI-based network visualization, which is accessible using the `influential::runShinyApp("IVI")` command. You can also access the shiny app online at https://influential.erc.monash.edu/.

ivi.from.indices(DC, CR, LH_index, NC, BC, CI, scaled = TRUE)

Arguments

DC

A vector containing the values of degree centrality of the desired vertices.

CR

A vector containing the values of ClusterRank of the desired vertices.

LH_index

A vector containing the values of local H-index of the desired vertices.

NC

A vector containing the values of neighborhood connectivity of the desired vertices.

BC

A vector containing the values of betweenness centrality of the desired vertices.

CI

A vector containing the values of Collective Influence of the desired vertices.

scaled

Logical; whether the end result should be 1-100 range normalized or not (default is TRUE).

Value

A numeric vector with the IVI values based on the provided centrality measures.

See also

Examples

MyData <- centrality.measures
My.vertices.IVI <- ivi.from.indices(DC = centrality.measures$DC,
                                    CR = centrality.measures$CR,
                                    NC = centrality.measures$NC,
                                    LH_index = centrality.measures$LH_index,
                                    BC = centrality.measures$BC,
                                    CI = centrality.measures$CI)